Handling AST building just before evaluate a user defined procedure.
Logo has dynamic scoping, so we are not able to detect if a symbol is available or not during procedure parsing.
So, we store original parsed statement and every time we have to evalute it, we build AST, check for symbols and
number of arguments of procedures and run the just created AST.

Added short form for print procedure (used in castle.logo)

Added method one to interpreter for parse and run just one string.

Added clear method to memory::Frame to erase every symbol defined and
start as a new frame.

These changes made examples/castle.logo executable.